Create custom useColorPicker hook
