+++ src/Components/ColorPickerPopper/index.js
@@ -1,3 +1,5 @@
 import ColorPickerPopper from "./ColorPickerPopper";
 
 export default ColorPickerPopper;
+
+export { useColorPicker } from "./useColorPicker";
